Break up util.py.
