+++ streaming/stream.py
@@ -18,8 +18,8 @@ from streaming.constant import TICK
 from streaming.distributed import barrier, get_local_rank
 from streaming.format import FileInfo, Reader, get_index_basename, reader_from_json
 from streaming.hashing import get_hash
-from streaming.storage import download_file
-from streaming.util import retry, wait_for_file_to_exist
+from streaming.storage import download_file, wait_for_file_to_exist
+from streaming.util import retry
 from streaming.world import World
